Handle multiple datasets with different contexts and XML file provided through http
For example, if BigDataViewer uses http://localhost:8080/HisYFP-SPIM, it will process the xml file directly and access the remote BigDataServer.
